Create 8-BSR
This repository contains a C program that simulates an 8-bit shift register, a fundamental concept in electronics and digital systems. 
The program allows users to:
-> Initialize the shift register to zero.
-> Perform left and right shift operations with user-defined input bits.
-> Simulate an 8-bit binary counter, showcasing all possible states sequentially.

The menu-driven interface makes it user-friendly for learning and experimentation Ideal for electronics enthusiasts, students, and developers looking to explore digital circuit simulations using C programming.